Separate component EnhancedTableHead from ChampionList
