+++ client/src/components/EnhancedTableHead/index.js
@@ -1,0 +1,3 @@
+import EnhancedTableHead from './EnhancedTableHead';
+
+export default EnhancedTableHead;
